Add unit tests for AdoPipelineTriggerService branch policy and pull request validation
- Implemented tests for IsPipelineRequiredByBranchPolicy method to verify behavior when pipeline is required, not in policy, policy is disabled, and no build policies exist.
- Added tests for RewirePipelineToGitHub method to ensure triggers are preserved or enabled based on branch policy requirements.
- Created AdoPipelineTriggerServiceFactory to facilitate the creation of AdoPipelineTriggerService instances with dependency injection.
- Enhanced test coverage for CreateYamlControlledTriggers and HasPullRequestTrigger methods to validate trigger creation logic.
