+++ src/ado2gh/Commands/RewirePipeline/RewirePipelineCommandHandler.cs
@@ -9,11 +9,13 @@ public class RewirePipelineCommandHandler : ICommandHandler<RewirePipelineComman
 {
     private readonly OctoLogger _log;
     private readonly AdoApi _adoApi;
+    private readonly AdoPipelineTriggerService _pipelineTriggerService;
 
-    public RewirePipelineCommandHandler(OctoLogger log, AdoApi adoApi)
+    public RewirePipelineCommandHandler(OctoLogger log, AdoApi adoApi, AdoPipelineTriggerService pipelineTriggerService)
     {
         _log = log;
         _adoApi = adoApi;
+        _pipelineTriggerService = pipelineTriggerService;
     }
 
     public async Task Handle(RewirePipelineCommandArgs args)
@@ -25,9 +27,23 @@ public class RewirePipelineCommandHandler : ICommandHandler<RewirePipelineComman
 
         _log.LogInformation($"Rewiring Pipeline to GitHub repo...");
 
+        // Use AdoApi for simple API calls
         var adoPipelineId = await _adoApi.GetPipelineId(args.AdoOrg, args.AdoTeamProject, args.AdoPipeline);
         var (defaultBranch, clean, checkoutSubmodules, triggers) = await _adoApi.GetPipeline(args.AdoOrg, args.AdoTeamProject, adoPipelineId);
-        await _adoApi.ChangePipelineRepo(args.AdoOrg, args.AdoTeamProject, adoPipelineId, defaultBranch, clean, checkoutSubmodules, args.GithubOrg, args.GithubRepo, args.ServiceConnectionId, triggers, args.TargetApiUrl);
+
+        // Use the specialized service for complex trigger logic
+        await _pipelineTriggerService.RewirePipelineToGitHub(
+            args.AdoOrg,
+            args.AdoTeamProject,
+            adoPipelineId,
+            defaultBranch,
+            clean,
+            checkoutSubmodules,
+            args.GithubOrg,
+            args.GithubRepo,
+            args.ServiceConnectionId,
+            triggers,
+            args.TargetApiUrl);
 
         _log.LogSuccess("Successfully rewired pipeline");
     }
